Rename yang filename to <module>@<revision> format
The O-RAN Alliance publish its yang modules with the filename format "<module>.yang".
For development purposes a filename in format "<module>@<revision>.yang" has the advantage,
that decisions by NetConf-Client and NetConf-Servers can be taken about compatibility and
API version without parsing the yang module.
This script parse the yang module for the latest revision and to rename the file accordingly.

IssueID: OAM-163
Change-Id: Ib7fd9aa0103d106947cd38f51a211a61338e66a8
Signed-off-by: demx8as6 <martin.skorupski@highstreet-technologies.com>